requirements.txt
aiohttp
cryptography
fastapi
uvicorn
python-dotenv